requirements/requirements.txt
Django==3.0.1
djangorestframework==3.11.0
drf-nested-routers==0.91
psycopg2-binary==2.8.4
